--- src/main/java/Duke.java
@@ -1,10 +1,0 @@
-public class Duke {
-    public static void main(String[] args) {
-        System.out.println("-----------------------------------");
-        System.out.println("Hello! I'm Cro!");
-        System.out.println("What can I do for you?");
-        System.out.println("-----------------------------------");
-        System.out.println("Bye. Hope to see you again soon!");
-        System.out.println("-----------------------------------");
-    }
-}
